+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Notes
 
+## [[v1.6.0](https://github.com/edseto/map/releases/tag/1.6.0)]
+
+### Added
+- Add remove marker method [e3705c5](https://github.com/edseto/map/commit/e3705c53578e6fb735565162f3ca08c44b557198)
+- Add anchor to set coordinates on icon base point [c74ec3e](https://github.com/edseto/map/commit/c74ec3efb78532c4aa8d828fde2146dea70faab6 )
+
+### Fixed
+- Refactor default options to keep default mapOptions and avoid losing default attributes not present in incoming options [68f4753](https://github.com/edseto/map/commit/68f4753c83249ad12ea1384a421b266a2f2c5fa3)
+- Prevent zoom on scroll in mobile by setting the dragging option [cb01179](https://github.com/edseto/map/commit/cb01179a4bcccaa55a7341950c6884041eb1a4f5 )
+
 ## [[v1.5.0](https://github.com/edseto/map/releases/tag/1.5.0)]
 
 ### Added